@@ -1,18 +1,16 @@
-"use strict";
+import React from "react";
 
-define(["react"], function(React) {
-    var height = 304;
+var height = 304;
 
-    return function FixedCode(props) {
-        return (
-            <pre
-                id="fixedEditor"
-                className="editor"
-                data-editor-lang="js"
-                style={{ height: height + "px" }}
-            >
-                {props.values}
-            </pre>
-        );
-    };
-});
+export default function FixedCode(props) {
+    return (
+        <pre
+            id="fixedEditor"
+            className="editor"
+            data-editor-lang="js"
+            style={{ height: height + "px" }}
+        >
+            {props.values}
+        </pre>
+    );
+}
