@@ -1,28 +1,27 @@
-"use strict";
+import React from "react";
+import events from "./events";
 
-define(["react", "events"], function(React, events) {
-    return function Message(props) {
-        var isError = props.value.fatal || props.value.severity === 2;
+export default function Message(props) {
+    var isError = props.value.fatal || props.value.severity === 2;
 
-        return (
-            <button
-                className={"alert btn-block report-message-alert alert-" + (isError ? "danger" : "warning")}
-                title={props.value.message}
-                onClick={
-                    function() {
-                        events.trigger("showError", props.value.line, props.value.column);
-                    }
+    return (
+        <button
+            className={"alert btn-block report-message-alert alert-" + (isError ? "danger" : "warning")}
+            title={props.value.message}
+            onClick={
+                function() {
+                    events.trigger("showError", props.value.line, props.value.column);
                 }
-            >
-                {props.value.line}:{props.value.column} - {props.value.message}
-                {
-                    !props.value.fatal && [
-                        " (",
-                        <a key="ruleLink" href={"https://eslint.org/docs/rules/" + props.value.ruleId}>{props.value.ruleId}</a>,
-                        ")"
-                    ]
-                }
-            </button>
-        );
-    };
-});
+            }
+        >
+            {props.value.line}:{props.value.column} - {props.value.message}
+            {
+                !props.value.fatal && [
+                    " (",
+                    <a key="ruleLink" href={"https://eslint.org/docs/rules/" + props.value.ruleId}>{props.value.ruleId}</a>,
+                    ")"
+                ]
+            }
+        </button>
+    );
+}
